refector content script message handler
